backend/django/requirements.txt
Django>=3.0,<4.0
djangorestframework
PyJWT
requests
sentry_sdk

